Improved downloading and conversion experience

+++ scripts/download_import_cldr.py
@@ -3,6 +3,7 @@
 import os
 import sys
 import shutil
+import hashlib
 import zipfile
 import urllib
 import subprocess
@@ -10,7 +11,49 @@ import subprocess
 
 URL = 'http://unicode.org/Public/cldr/1.9.1/core.zip'
 FILENAME = 'core-1.9.1.zip'
+FILESUM = '1c506cd7a30bf5b4f3cbb8a5b382d96c'
 BLKSIZE = 131072
+
+
+def get_terminal_width():
+    import fcntl
+    import termios
+    import struct
+    fd = sys.stdin.fileno()
+    cr = struct.unpack('hh', fcntl.ioctl(fd, termios.TIOCGWINSZ, '1234'))
+    return cr[1]
+
+
+def reporthook(block_count, block_size, total_size):
+    bytes_transmitted = block_count * block_size
+    cols = get_terminal_width()
+    buffer = 6
+    percent = float(bytes_transmitted) / (total_size or 1)
+    done = int(percent * (cols - buffer))
+    sys.stdout.write('\r')
+    sys.stdout.write(' ' + '=' * done + ' ' * (cols - done - buffer))
+    sys.stdout.write('% 4d%%' % (percent * 100))
+    sys.stdout.flush()
+
+
+def log(message, *args):
+    if args:
+        message = message % args
+    print >> sys.stderr, message
+
+
+def is_good_file(filename):
+    if not os.path.isfile(filename):
+        log('\'%s\' not found', filename)
+        return False
+    h = hashlib.md5()
+    with open(filename, 'rb') as f:
+        while 1:
+            blk = f.read(BLKSIZE)
+            if not blk:
+                break
+            h.update(blk)
+        return h.hexdigest() == FILESUM
 
 
 def main():
@@ -18,24 +61,26 @@ def main():
     repo = os.path.dirname(scripts_path)
     cldr_path = os.path.join(repo, 'cldr')
     zip_path = os.path.join(cldr_path, FILENAME)
+    changed = False
 
-    if not os.path.isfile(zip_path):
-        with open(zip_path, 'wb') as f:
-            conn = urllib.urlopen(URL)
-            while True:
-                buf = conn.read(BLKSIZE)
-                if not buf:
-                    break
-                f.write(buf)
-            conn.close()
-
+    while not is_good_file(zip_path):
+        log('Downloading \'%s\'', FILENAME)
+        if os.path.isfile(zip_path):
+            os.remove(zip_path)
+        urllib.urlretrieve(URL, zip_path, reporthook)
+        changed = True
+        print
     common_path = os.path.join(cldr_path, 'common')
-    if os.path.isdir(common_path):
-        shutil.rmtree(common_path)
 
-    z = zipfile.ZipFile(zip_path)
-    z.extractall(cldr_path)
-    z.close()
+    if changed:
+        if os.path.isdir(common_path):
+            log('Deleting old CLDR checkout in \'%s\'', cldr_path)
+            shutil.rmtree(common_path)
+
+        log('Extracting CLDR to \'%s\'', cldr_path)
+        z = zipfile.ZipFile(zip_path)
+        z.extractall(cldr_path)
+        z.close()
 
     subprocess.check_call([
         sys.executable,
